+++ grails-app/services/io/xh/hoist/log/LogLevelService.groovy
@@ -12,10 +12,9 @@ import io.xh.hoist.BaseService
 import ch.qos.logback.classic.Logger
 import ch.qos.logback.classic.LoggerContext
 import ch.qos.logback.classic.Level
+import io.xh.hoist.cluster.ClusterRequest
 import io.xh.hoist.util.Utils
 import org.slf4j.LoggerFactory
-
-import java.util.concurrent.Callable
 
 import static io.xh.hoist.util.DateTimeUtils.MINUTES
 
@@ -85,10 +84,9 @@ class LogLevelService extends BaseService {
     void noteLogLevelChanged() {
         clusterService.submitToAllInstances (new CalculateAdjustments())
     }
-    static class CalculateAdjustments implements Callable, Serializable {
-        Map call() {
+    static class CalculateAdjustments extends ClusterRequest<Void> {
+        def doCall() {
             Utils.appContext.logLevelService.calculateAdjustments()
-            return [success: true]
         }
     }