+++ src/main/groovy/io/xh/hoist/cluster/ClusterResponse.groovy
@@ -1,9 +1,8 @@
 package io.xh.hoist.cluster
 
-class ClusterResponse implements Serializable {
-
-    int status
-    String result
+class ClusterResponse<T> implements Serializable {
+    T value
+    Throwable exception
 }